+++ libs/base/include/mrpt/math/data_utils.h
@@ -113,7 +113,7 @@ namespace mrpt
 		mahalanobisDistance2(const VECTORLIKE &delta_mu,const MATRIXLIKE &cov)
 		{
 			ASSERTDEB_(cov.isSquare())
-			ASSERTDEB_(cov.getColCount()==delta_mu.size())
+			ASSERTDEB_(size_t(cov.getColCount())==size_t(delta_mu.size()))
 			return multiply_HCHt_scalar(delta_mu,cov.inverse());
 		}
